+++ AsyncListenerList.js
@@ -1,63 +1,63 @@
 "use strict";
 var Q = require("q");
-var EventListenerList = require("./EventListenerList");
+var ListenerList = require("./ListenerList");
 /** An event listener list for asynchronous event listeners (i.e. the listeners perform asynchronous operations)
  * manages a list of listener functions and allows events to be sent to the listeners
  */
-var AsyncEventListenerList = (function () {
-    function AsyncEventListenerList() {
-        this.eventHandler = new EventListenerList();
+var AsyncListenerList = (function () {
+    function AsyncListenerList() {
+        this.eventHandler = new ListenerList();
     }
-    AsyncEventListenerList.prototype.reset = function () {
+    AsyncListenerList.prototype.reset = function () {
         this.eventHandler.reset();
     };
     // have to proxy all the methods because TypeScript 'extends ...' doesn't work with our gulp compilation process
-    AsyncEventListenerList.prototype.getListeners = function () {
+    AsyncListenerList.prototype.getListeners = function () {
         return this.eventHandler.getListeners();
     };
-    AsyncEventListenerList.prototype.getFireEventsSuccessCallback = function () {
+    AsyncListenerList.prototype.getFireEventsSuccessCallback = function () {
         return this.eventHandler.getFireEventsSuccessCallback();
     };
-    AsyncEventListenerList.prototype.setFireEventsSuccessCallback = function (cb) {
+    AsyncListenerList.prototype.setFireEventsSuccessCallback = function (cb) {
         this.eventHandler.setFireEventsSuccessCallback(cb);
     };
-    AsyncEventListenerList.prototype.getFireEventsFailureCallback = function () {
+    AsyncListenerList.prototype.getFireEventsFailureCallback = function () {
         return this.eventHandler.getFireEventsFailureCallback();
     };
-    AsyncEventListenerList.prototype.setFireEventsFailureCallback = function (cb) {
+    AsyncListenerList.prototype.setFireEventsFailureCallback = function (cb) {
         this.eventHandler.setFireEventsFailureCallback(cb);
     };
-    AsyncEventListenerList.prototype.getListenerAddedCallback = function () {
+    AsyncListenerList.prototype.getListenerAddedCallback = function () {
         return this.eventHandler.getListenerAddedCallback();
     };
-    AsyncEventListenerList.prototype.setListenerAddedCallback = function (cb) {
+    AsyncListenerList.prototype.setListenerAddedCallback = function (cb) {
         this.eventHandler.setListenerAddedCallback(cb);
     };
-    AsyncEventListenerList.prototype.getListenerRemovedCallback = function () {
+    AsyncListenerList.prototype.getListenerRemovedCallback = function () {
         return this.eventHandler.getListenerRemovedCallback();
     };
-    AsyncEventListenerList.prototype.setListenerRemovedCallback = function (cb) {
+    AsyncListenerList.prototype.setListenerRemovedCallback = function (cb) {
         this.eventHandler.setListenerRemovedCallback(cb);
     };
-    AsyncEventListenerList.prototype.addListener = function (listener) {
+    AsyncListenerList.prototype.addListener = function (listener) {
         this.eventHandler.addListener(listener);
     };
-    AsyncEventListenerList.prototype.addOneTimeListener = function (listener) {
+    AsyncListenerList.prototype.addOneTimeListener = function (listener) {
         this.eventHandler.addOneTimeListener(listener);
     };
-    AsyncEventListenerList.prototype.addNTimeListener = function (removeAfterNCalls, listener) {
+    AsyncListenerList.prototype.addNTimeListener = function (removeAfterNCalls, listener) {
         this.eventHandler.addNTimeListener(removeAfterNCalls, listener);
     };
-    AsyncEventListenerList.prototype.removeListener = function (listener) {
+    AsyncListenerList.prototype.removeListener = function (listener) {
         this.eventHandler.removeListener(listener);
     };
-    AsyncEventListenerList.prototype.removeListenerAt = function (index) {
+    AsyncListenerList.prototype.removeListenerAt = function (index) {
         this.eventHandler.removeListenerAt(index);
     };
     /**
      * @param event: the event object being fired to listeners
      */
-    AsyncEventListenerList.prototype.fireEvent = function (event, customListenerCaller, customListenerCallsDoneCb) {
+    AsyncListenerList.prototype.fireEvent = function (event, customListenerCaller, customListenerCallsDoneCb) {
         var fireEventsSuccessCallback = this.getFireEventsSuccessCallback();
         var fireEventsFailureCallback = this.getFireEventsFailureCallback();
         var dfds = [];
@@ -80,6 +80,6 @@ var AsyncEventListenerList = (function () {
             });
         });
     };
-    return AsyncEventListenerList;
+    return AsyncListenerList;
 }());
-module.exports = AsyncEventListenerList;
+module.exports = AsyncListenerList;