+++ test/ListenerListTest.js
@@ -1,45 +1,33 @@
-﻿import chai = require("chai");
-import mocha = require("mocha");
-import EventListenerList = require("../EventListenerList");
-
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var chai = require("chai");
+var ListenerList = require("../ListenerList");
 var asr = chai.assert;
-
-
-suite("EventListenerList", function ArraysTest() {
-
-    var createListenerList = () => EventListenerList.newInst<{ chance: number; name: string }>();
-
-    var sortNames = (a, b) => (<string>a.name).localeCompare(b.name);
-
-
+suite("ListenerList", function ListenerListTest() {
+    var createListenerList = function () { return ListenerList.newInst(); };
+    var sortNames = function (a, b) { return a.name.localeCompare(b.name); };
     test("addListenerAndFireEvents", function addListenerAndFireEventsTest() {
         var ell = createListenerList();
         var all = 0;
         var once = 0;
         var twice = 0;
         var thrice = 0;
-        ell.addListener(() => all++);
-        ell.addOneTimeListener(() => once++);
-        ell.addNTimeListener(2, () => twice++);
-        ell.addNTimeListener(3, () => thrice++);
-
+        ell.addListener(function () { return all++; });
+        ell.addOneTimeListener(function () { return once++; });
+        ell.addNTimeListener(2, function () { return twice++; });
+        ell.addNTimeListener(3, function () { return thrice++; });
         ell.fireEvent({ name: "one", chance: 0.0 });
         ell.fireEvent({ name: "two", chance: 0.2 });
         ell.fireEvent({ name: "three", chance: 0.4 });
         ell.fireEvent({ name: "four", chance: 0.6 });
         ell.fireEvent({ name: "five", chance: 0.8 });
-
         ell.reset();
-
         ell.fireEvent({ name: "six", chance: 1 });
-
         asr.equal(all, 5);
         asr.equal(once, 1);
         asr.equal(twice, 2);
         asr.equal(thrice, 3);
     });
-
-
     test("addAndRemoveListener", function addAndRemoveListenerTest() {
         var ell = createListenerList();
         var added = 0;
@@ -50,40 +38,28 @@ suite("EventListenerList", function ArraysTest() {
         var lstA = function a() { aC++; };
         var lstB = function b() { bC++; };
         var lstC = function c() { cC++; };
-
-        ell.setListenerAddedCallback(() => added++);
-        ell.setListenerRemovedCallback(() => removed++);
-
+        ell.setListenerAddedCallback(function () { return added++; });
+        ell.setListenerRemovedCallback(function () { return removed++; });
         ell.addListener(lstA);
         ell.addListener(lstB);
-
         asr.deepEqual(ell.getListeners().sort(sortNames), [lstA, lstB]);
-
         ell.removeListener(lstA);
         ell.addListener(lstC);
-
         asr.deepEqual(ell.getListeners().sort(sortNames), [lstB, lstC]);
-
         asr.equal(added, 3);
         asr.equal(removed, 1);
     });
-
-
     test("fireEvents", function fireEventsTest() {
         var ell = createListenerList();
         var success = 0;
         var failure = 0;
-        ell.setFireEventsSuccessCallback(() => success++);
-        ell.setFireEventsFailureCallback(() => failure++);
-
-        ell.addListener((evt) => evt.chance);
-
+        ell.setFireEventsSuccessCallback(function () { return success++; });
+        ell.setFireEventsFailureCallback(function () { return failure++; });
+        ell.addListener(function (evt) { return evt.chance; });
         ell.fireEvent({ name: "1", chance: 0.1 });
         ell.fireEvent({ name: "3", chance: 0.3 });
-        ell.fireEvent({ name: "1", get chance(): number { throw new Error("event 3 failure"); } });
-
+        ell.fireEvent({ name: "1", get chance() { throw new Error("event 3 failure"); } });
         asr.equal(success, 2);
         asr.equal(failure, 1);
-    })
-
+    });
 });